+++ Makefile
@@ -14,7 +14,7 @@ SOURCES =		minishell.c	main.c signals.c heredoc.c exit_minishell.c init.c heredo
 LEXER =			lexer.c handle_spaces.c
 LIST =			ft_lstadd_back_t.c ft_lstnew_t.c ft_lstfoward_free_t.c \
 				ft_lstsize_tab.c ft_lstclear_t.c
-PARSER =		parser.c expander.c parser_error.c expander_2.c
+PARSER =		parser.c expander.c parser_error.c expander_2.c test_open.c
 EXECUTOR =		exec_clean.c pipe.c redirects.c executor.c exec_error.c
 BUILT_IN =		is_built_in.c echo.c cd.c pwd.c built_in_error.c unset.c \
 				exportation.c exportation_2.c env.c exit_builtin.c perror_message.c
