+++ inc/minishell.h
@@ -6,7 +6,7 @@
 /*   By: ralves-b <ralves-b@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/16 23:11:52 by wportilh          #+#    #+#             */
-/*   Updated: 2022/10/03 16:51:40 by ralves-b         ###   ########.fr       */
+/*   Updated: 2022/10/03 18:28:28 by ralves-b         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -165,7 +165,7 @@ void		lexer(t_tokens **tks, char **str, t_table **tab);
  * @param tks 
  * @param tab 
  */
-void		parser(t_tokens **tks, t_table *tab);
+void		parser(t_tokens **tks, t_table **tab);
 
 /**
  * @brief init variables next and envp of the table node
@@ -254,5 +254,6 @@ void		false_all(t_table *tab);
 void		reset_tab(char *buff);
 void		prepare_infile(t_tokens **tks, t_table **tab);
 void		check_heredoc(void);
+void		teste_open(int red, char **file, t_table **tab);
 
 #endif
