+++ src/executor/executor.c
@@ -6,7 +6,7 @@
 /*   By: ralves-b <ralves-b@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/19 18:47:12 by wportilh          #+#    #+#             */
-/*   Updated: 2022/09/30 18:26:51 by ralves-b         ###   ########.fr       */
+/*   Updated: 2022/10/03 19:13:38 by ralves-b         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
